refactor: 코드 분리
- 상수 constants/location.ts로 분리
- 지도 렌더링 컴포넌트 runMap.tsx로 분리
- 위치 추적 코드 hooks/useLocationTracking.ts로 분리
- haversineDistance 함수 utilst/location.ts로 분리
