+++ src/constants/location.ts
@@ -1,0 +1,4 @@
+export const LOCATION_UPDATE_INTERVAL_MS = 1000;
+export const LOCATION_DISTANCE_INTERVAL_M = 10;
+export const MIN_DISTANCE_THRESHOLD_M = 10;
+export const INITIAL_ZOOM_LEVEL = 14;
