+++ src/utils/location.ts
@@ -1,0 +1,30 @@
+import type { Position } from 'geojson';
+
+const EARTH_RADIUS_IN_METERS = 6371e3;
+
+export const haversineDistance = (
+  coords1: Position,
+  coords2: Position,
+): number => {
+  const toRad = (x: number) => (x * Math.PI) / 180;
+
+  const lon1 = coords1[0];
+  const lat1 = coords1[1];
+  const lon2 = coords2[0];
+  const lat2 = coords2[1];
+
+  const R = EARTH_RADIUS_IN_METERS;
+
+  const dLat = toRad(lat2 - lat1);
+  const dLon = toRad(lon2 - lon1);
+  const a =
+    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+    Math.cos(toRad(lat1)) *
+      Math.cos(toRad(lat2)) *
+      Math.sin(dLon / 2) *
+      Math.sin(dLon / 2);
+  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+  const d = R * c;
+
+  return d;
+};
